Add files via upload
Lab automation system will control lightings, fans, entertainment systems, and appliances. It also include door security such as alarm systems. When connected with the Internet, devices are an important constituent of the Internet of Things ("IoT").

The lab automation system typically connects controlled devices to a central hub or "gateway". The user interface for control of the system uses either wall-mounted terminals, tablet or desktop computers, a mobile phone application, or a Web interface, that may also be accessible off-site through the Internet.

Basically, the concept of lab automation works by connecting smart devices and sensors (wired or wireless) to a system which act as the main control, and lastly to a graphical user interface for examples for smart devices/sensors are like humidity sensors, proximity sensors, brightness sensors, etc.